crawler/requirements.crawler.txt
psycopg2-binary>=2.9.9
requests>=2.31.0
lxml>=4.9.3
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
PyYAML>=6.0.1
pythainlp>=4.0.2
tinysegmenter==0.3
tldextract>=5.1.2
cssselect>=1.2.0
feedparser>=6.0.10
feedfinder2>=0.0.4
jieba3k>=0.35.1
Pillow>=10.4.0

